DocuMind-RAG-Assistant/requirements.txt
langchain==1.0.3
langchain-groq==1.0.0
langchain-huggingface==1.0.1
langchain-chroma==1.0.0
langchain-community==0.4.1
langchain-text-splitters==1.0.0

python-dotenv==1.2.1
chromadb==1.3.4
sentence-transformers==5.1.2
huggingface-hub==0.25.1

openai==2.7.1
groq==0.33.0
pydantic==2.12.4
